requirements.txt
# TripCompare API Dependencies

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# HTTP client for external APIs
httpx==0.26.0
aiohttp==3.9.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==23.12.1
isort==5.13.2
